refactor: move shutdownhook releated code from BrokerController to BrokerServiceManager
